// next/lib/services/models/article.ts
/* tslint:disable */
/* eslint-disable */
/**
 * DOCUMENTATION
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: contact-email@something.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ArticleCategoriesInner } from './article-categories-inner';
// May contain unused imports in some cases
// @ts-ignore
import type { ArticleCategoriesInnerProductImagesInner } from './article-categories-inner-product-images-inner';
// May contain unused imports in some cases
// @ts-ignore
import type { ArticleCategoriesInnerProductImagesInnerRelatedInner } from './article-categories-inner-product-images-inner-related-inner';
// May contain unused imports in some cases
// @ts-ignore
import type { ArticleRequestDataDynamicZoneInner } from './article-request-data-dynamic-zone-inner';
// May contain unused imports in some cases
// @ts-ignore
import type { SharedSeoComponent } from './shared-seo-component';

/**
 * 
 * @export
 * @interface Article
 */
export interface Article {
    /**
     * 
     * @type {number}
     * @memberof Article
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof Article
     */
    'documentId'?: string;
    /**
     * 
     * @type {SharedSeoComponent}
     * @memberof Article
     */
    'seo'?: SharedSeoComponent;
    /**
     * 
     * @type {string}
     * @memberof Article
     */
    'title'?: string;
    /**
     * 
     * @type {string}
     * @memberof Article
     */
    'description'?: string;
    /**
     * 
     * @type {string}
     * @memberof Article
     */
    'slug'?: string;
    /**
     * 
     * @type {any}
     * @memberof Article
     */
    'content'?: any;
    /**
     * 
     * @type {Array<ArticleCategoriesInner>}
     * @memberof Article
     */
    'categories'?: Array<ArticleCategoriesInner>;
    /**
     * 
     * @type {Array<ArticleRequestDataDynamicZoneInner>}
     * @memberof Article
     */
    'dynamic_zone'?: Array<ArticleRequestDataDynamicZoneInner>;
    /**
     * 
     * @type {ArticleCategoriesInnerProductImagesInner}
     * @memberof Article
     */
    'image'?: ArticleCategoriesInnerProductImagesInner;
    /**
     * 
     * @type {string}
     * @memberof Article
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof Article
     */
    'updatedAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof Article
     */
    'publishedAt'?: string;
    /**
     * 
     * @type {ArticleCategoriesInnerProductImagesInnerRelatedInner}
     * @memberof Article
     */
    'createdBy'?: ArticleCategoriesInnerProductImagesInnerRelatedInner;
    /**
     * 
     * @type {ArticleCategoriesInnerProductImagesInnerRelatedInner}
     * @memberof Article
     */
    'updatedBy'?: ArticleCategoriesInnerProductImagesInnerRelatedInner;
    /**
     * 
     * @type {string}
     * @memberof Article
     */
    'locale'?: string;
    /**
     * 
     * @type {Array<ArticleCategoriesInnerProductImagesInnerRelatedInner>}
     * @memberof Article
     */
    'localizations'?: Array<ArticleCategoriesInnerProductImagesInnerRelatedInner>;
}